app/requirements.txt
yfinance
pandas
numpy
openpyxl
PyYAML
streamlit
plotly